requirements.txt
fastapi>=0.115.12,<0.116.0
uvicorn>=0.34.0,<0.35.0
pymongo[srv]>=4.13.0,<5.0.0
passlib[bcrypt]>=1.7.4,<2.0.0
python-jose[cryptography]>=3.5.0,<4.0.0
python-multipart>=0.0.20,<0.0.21
python-dotenv>=1.1.0,<2.0.0
bcrypt==4.0.1
openai>=1.0.0,<2.0.0
pydantic[email]>=2.11.7,<3.0.0
email-validator>=2.0.0,<3.0.0
google-api-python-client>=2.0.0,<3.0.0
google-auth-httplib2>=0.1.0,<1.0.0
google-auth-oauthlib>=1.0.0,<2.0.0
google-auth>=2.0.0,<3.0.0
cryptography>=3.0.0,<4.0.0
sendgrid>=6.12.4,<7.0.0
twilio>=8.0.0,<9.0.0

# Testing
pytest>=8.0,<9.0
pytest-asyncio>=0.24,<1.0
httpx>=0.27,<1.0
pytest-cov>=5.0,<6.0